move __version__ from _version.py to version.py
