+++ pastas/__init__.py
@@ -8,7 +8,6 @@ import pastas.stats as stats
 import pastas.timeseries_utils as ts
 
 
-from ._version import __version__
 from .decorators import set_use_numba
 from .model import Model
 from .modelcompare import CompareModels
@@ -41,7 +40,7 @@ from .stressmodels import (
 from .timeseries import validate_oseries, validate_stress
 from .transform import ThresholdTransform
 from .utils import initialize_logger, set_log_level
-from .version import show_versions
+from .version import show_versions, __version__
 
 logger = logging.getLogger(__name__)
 initialize_logger(logger)
