+++ pyproject.toml
@@ -72,7 +72,7 @@ numbascipy = ["numba-scipy >= 0.3.1"]
 
 
 [tool.setuptools.dynamic]
-version = { attr = "pastas._version.__version__" }
+version = { attr = "pastas.version.__version__" }
 
 [tool.black]
 line-length = 88
